+++ psydac/feec/multipatch/examples/hcurl_source_pbms_conga_2d.py
@@ -1,3 +1,5 @@
+# coding: utf-8
+
 from mpi4py import MPI
 
 import os
@@ -301,26 +303,6 @@ def solve_hcurl_source_pbm(
         uh_c      = array_to_stencil(uh_c, V1h.vector_space)
         l2_error  = l2norm_h.assemble(u=FemField(V1h, coeffs=uh_c))
         return l2_error
-
-def test_time_harmonic_maxwell_pretzel_f():
-    nc,deg      = 10,2
-    source_type = 'manu_maxwell'
-    domain_name = 'pretzel_f'
-
-    omega = np.sqrt(170) # source
-    roundoff = 1e4
-    eta = int(-omega**2 * roundoff)/roundoff
-
-    l2_error = solve_hcurl_source_pbm(
-        nc=nc, deg=deg,
-        eta=eta,
-        nu=0,
-        mu=1, #1,
-        domain_name=domain_name,
-        source_type=source_type,
-        backend_language='pyccel-gcc')
-
-    assert abs(l2_error - 0.06246693595198972)<1e-10
 
 if __name__ == '__main__':
